ISSUE #2087 - Move common updateButton component
