+++ frontend/routes/viewerGui/components/risks/components/riskDetails/riskDetails.styles.ts
@@ -24,29 +24,9 @@ import { cond, constant, matches, stubTrue } from 'lodash';
 import { COLOR } from '../../../../../../styles';
 import { MessagesList as MessagesListComponent } from '../../../../../components/messagesList/';
 import * as TextFieldStyles from '../../../../../components/textField/textField.styles';
-import { FileUploadInvoker } from '../../../commentForm/commentForm.styles';
-import { StyledButton } from '../../../containedButton/containedButton.styles';
-import { Container as ButtonContainer } from '../../../pinButton/pinButton.styles';
 import PreviewDetailsBase from '../../../previewDetails/previewDetails.container';
 
 export const StyledFormControl = styled(FormControl)`
-`;
-
-export const UpdateButtonsContainer = styled(FormControl)`
-	&& {
-		display: flex;
-		flex-direction: row;
-		justify-content: flex-start;
-		width: 100%;
-
-		${StyledButton} {
-			padding: 4px 10px;
-		}
-
-		${ButtonContainer} ~ ${ButtonContainer} {
-			margin-left: 20px;
-		}
-	}
 `;
 
 export const MessagesList = styled(MessagesListComponent)`
